requirements.txt
dbt-bigquery==1.7.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.0
pandas==2.2.0
